chore: comprehensive code quality improvements and test fixes
This commit includes systematic code quality improvements across the entire codebase:

## Compilation Fixes
- Fix DNS cache API usage in crates/sb-core/tests/dns_steady.rs
  - Update cache key from &str to proper Key struct
  - Add QType::A parameter for cache operations

## Code Quality Improvements
- Remove unused imports across test files:
  - app/tests/prometheus_integration_robustness.rs
  - app/tests/admin_observe.rs (2 locations)
  - crates/sb-core/tests/router_rules_port_transport.rs

## Verification
- All tests now compile successfully
- cargo clippy --all-features --workspace passes with -D warnings
- cargo build --all-features --workspace completes without errors

## Code Scan Results
- No TODO/FIXME/XXX markers found in production code
- All panic!/unreachable! usage verified (tests only)
- unwrap/expect usage reviewed (1580 total, primarily in tests)
- Zero clippy warnings with strict mode

+++ crates/sb-core/src/router/cache_hot.rs
@@ -64,7 +64,7 @@ pub fn hot_json(limit: usize) -> String {
     for it in items {
         parts.push(minijson::obj([
             ("hash_prefix", minijson::Val::Str(&it.hash_prefix)),
-            ("hits", minijson::Val::NumU(it.hits as u64)),
+            ("hits", minijson::Val::NumU(it.hits)),
         ]));
     }
     let arr = format!("[{}]", parts.join(","));
